New features: truncate individual radials; threshold SNR
Add ability to truncate individual radials, so now the study area need not be a perfect circle. As part of this feature the monte-carlo code for calculating p(det) can handle NA/NaN occurrance in TL data.

Additionally, add an SNR threshold to the monte-carlo simulation to force SNR below this threshold to be zero. This can be useful for GAMs that don't behave, but especially for VGAMs where the resulting function asymptotes to a non-zero value with decreasing SNR.